pyproject.toml
[project]
name = "codeius"
version = "1.0.50"
description = "AI-powered coding assistant with multiple tools and visual interface"
readme = "README.md"
license = { text = "MIT" }
authors = [
    { name = "MuhammadUsmanGM", email = "MuhammadUsmanGM@users.noreply.github.com" }
]
requires-python = ">=3.11"
dependencies = [
    "gitpython>=3.1.45",
    "prompt-toolkit>=3.0.52",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "rich>=14.2.0",
    "pyfiglet>=1.0.0",
    "pydantic>=2.12.4",
    "flask>=3.1.2",
    "pytest>=9.0.0",
    "pillow>=12.0.0",
    "pytesseract>=0.3.13",
    "radon>=6.0.1",
    "flake8>=7.3.0",
    "networkx>=3.2.1",
    "matplotlib>=3.8.0",
    "pyyaml>=6.0",
    "build>=1.3.0",
    "twine>=6.2.0",
    "cachetools>=5.3.0",
    "flask-socketio>=5.3.6",
    "flask-cors>=5.0.0",
    "keyboard>=0.13.5",
    "flask-login>=0.6.3",
    "authlib>=1.6.5",
    "pymongo>=4.15.4",
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Topic :: Software Development :: Build Tools",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
]

[project.scripts]
codeius = "codeius.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/codeius"]
